refactor class add auth database #2
