+++ src/test/java/com/sc2/hackathome/HackAtHomeBeApplicationTests.java
@@ -1,4 +1,4 @@
-package com.sc2.hackathome.java;
+package com.sc2.hackathome;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
